moved get_seed_id tests
