+++ tests/test_events/test_catalog_utils.py
@@ -437,3 +437,34 @@ class TestMakeOrigins:
         ori = strange_picks_added_origins[0].origins[0]
         time = strange_picks_added_origins[1].time
         assert ori.time == time
+
+
+class TestGetSeedId:
+    """Tests for the get_seed_id function"""
+
+    def test_get_seed_id(self):
+        """Make sure it is possible to retrieve the seed id"""
+        wid = obspy.core.event.WaveformStreamID(
+            network_code="AA",
+            station_code="BBB",
+            location_code="CC",
+            channel_code="DDD",
+        )
+        pick = obspy.core.event.Pick(waveform_id=wid)
+        amp = obspy.core.event.Amplitude(pick_id=pick.resource_id)
+        station_mag = obspy.core.event.StationMagnitude(amplitude_id=amp.resource_id)
+        station_mag_cont = obspy.core.event.StationMagnitudeContribution(
+            station_magnitude_id=station_mag.resource_id
+        )
+        seed = obsplus.events.utils.get_seed_id(station_mag_cont)
+        assert seed == "AA.BBB.CC.DDD"
+
+    def test_no_seed_id(self):
+        """Make sure raises AttributeError if no seed info found"""
+        with pytest.raises(AttributeError):
+            obsplus.events.utils.get_seed_id(obspy.core.event.Pick())
+
+    def test_unsupported(self):
+        """Make sure an unsupported object raises TypeError"""
+        with pytest.raises(TypeError):
+            obsplus.events.utils.get_seed_id(obspy.core.event.Origin())
